# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend server for the UTT Report Maintenance application.",
        "title": "UTT Report Maintenance API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "report-utt.web.app",
    "basePath": "/",
    "paths": {
        "/api/report": {
            "post": {
                "description": "Save a new report document to the specified collection.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Create a new report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Secret",
                        "name": "X-API-Secret",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Firebase JWT Token (Bearer)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Report data",
                        "name": "report",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateReportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/report/{collection}/{id}": {
            "get": {
                "description": "Retrieve a single report document from a specific collection.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Get a report by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Secret",
                        "name": "X-API-Secret",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Firebase JWT Token (Bearer)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Collection name (e.g., hse)",
                        "name": "collection",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Report ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a report document from the specified collection.",
                "tags": [
                    "reports"
                ],
                "summary": "Delete a report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Secret",
                        "name": "X-API-Secret",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Firebase JWT Token (Bearer)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Collection name",
                        "name": "collection",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Report ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/reports": {
            "get": {
                "description": "Retrieve a list of reports from a specific collection.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "List reports in a collection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Secret",
                        "name": "X-API-Secret",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Firebase JWT Token (Bearer)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Collection name (defaults to hse)",
                        "name": "collection",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.SuccessResponse"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CreateReportRequest": {
            "type": "object",
            "required": [
                "collection",
                "report_type",
                "title"
            ],
            "properties": {
                "collection": {
                    "type": "string",
                    "enum": [
                        "hse",
                        "pdf_documents",
                        "excel_documents",
                        "service_reports"
                    ]
                },
                "description": {
                    "type": "string",
                    "maxLength": 500
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Photo"
                    }
                },
                "report_type": {
                    "$ref": "#/definitions/models.ReportType"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.Photo": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "caption": {
                    "type": "string",
                    "maxLength": 200
                },
                "storage_path": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.ReportType": {
            "type": "string",
            "enum": [
                "hse",
                "pdf_documents",
                "excel_documents",
                "service_reports"
            ],
            "x-enum-varnames": [
                "ReportTypeHSE",
                "ReportTypePDF",
                "ReportTypeExcel",
                "ReportTypeService"
            ]
        },
        "models.SuccessResponse": {
            "type": "object",
            "properties": {
                "collection": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "reportId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}